Converted Test from TextAndGrpahicsManagerTest to TextAndGraphicUpdateOperationTest
