+++ android/sdl_android/src/androidTest/java/com/smartdevicelink/managers/screen/TextAndGraphicManagerTests.java
@@ -154,494 +154,6 @@ public class TextAndGraphicManagerTests {
 		assertEquals(ManagerUtility.WindowCapabilityUtility.getMaxNumberOfMainFieldLines(textAndGraphicManager.defaultMainWindowCapability), 3);
 	}
 
-/*	@Test
-	public void testAssemble1Line(){
-
-		Show inputShow = new Show();
-
-		// Force it to return display with support for only 1 line of text
-		textAndGraphicManager.defaultMainWindowCapability = getWindowCapability(1);
-
-		textAndGraphicManager.setTextField1("It is");
-		textAndGraphicManager.setTextField1Type(MetadataType.HUMIDITY);
-
-		Show assembledShow = textAndGraphicManager.assembleShowText(inputShow);
-		assertEquals(assembledShow.getMainField1(), "It is");
-
-		// test tags (just 1)
-		MetadataTags tags = assembledShow.getMetadataTags();
-		List<MetadataType> tagsList = new ArrayList<>();
-		tagsList.add(MetadataType.HUMIDITY);
-		assertEquals(tags.getMainField1(), tagsList);
-
-		textAndGraphicManager.setTextField2("Wednesday");
-
-		assembledShow = textAndGraphicManager.assembleShowText(inputShow);
-		assertEquals(assembledShow.getMainField1(), "It is - Wednesday");
-
-		textAndGraphicManager.setTextField3("My");
-
-		assembledShow = textAndGraphicManager.assembleShowText(inputShow);
-		assertEquals(assembledShow.getMainField1(), "It is - Wednesday - My");
-
-		textAndGraphicManager.setTextField4("Dudes");
-		textAndGraphicManager.setTextField4Type(MetadataType.CURRENT_TEMPERATURE);
-
-		assembledShow = textAndGraphicManager.assembleShowText(inputShow);
-		assertEquals(assembledShow.getMainField1(), "It is - Wednesday - My - Dudes");
-
-		// test tags
-		tags = assembledShow.getMetadataTags();
-		tagsList = new ArrayList<>();
-		tagsList.add(MetadataType.HUMIDITY);
-		tagsList.add(MetadataType.CURRENT_TEMPERATURE);
-		assertEquals(tags.getMainField1(), tagsList);
-
-		// For some obscurity, lets try setting just fields 2 and 4 for a 1 line display
-		textAndGraphicManager.setTextField1(null);
-		textAndGraphicManager.setTextField3(null);
-
-		assembledShow = textAndGraphicManager.assembleShowText(inputShow);
-		assertEquals(assembledShow.getMainField1(), "Wednesday - Dudes");
-	}
-
-	@Test
-	public void testAssemble2Lines() {
-
-		Show inputShow = new Show();
-
-		// Force it to return display with support for only 2 lines of text
-		textAndGraphicManager.defaultMainWindowCapability = getWindowCapability(2);
-
-		textAndGraphicManager.setTextField1("It is");
-		textAndGraphicManager.setTextField1Type(MetadataType.HUMIDITY);
-
-		Show assembledShow = textAndGraphicManager.assembleShowText(inputShow);
-		assertEquals(assembledShow.getMainField1(), "It is");
-
-		// test tags
-		MetadataTags tags = assembledShow.getMetadataTags();
-		List<MetadataType> tagsList = new ArrayList<>();
-		tagsList.add(MetadataType.HUMIDITY);
-		assertEquals(tags.getMainField1(), tagsList);
-
-		textAndGraphicManager.setTextField2("Wednesday");
-		textAndGraphicManager.setTextField2Type(MetadataType.CURRENT_TEMPERATURE);
-
-		assembledShow = textAndGraphicManager.assembleShowText(inputShow);
-		assertEquals(assembledShow.getMainField1(), "It is");
-		assertEquals(assembledShow.getMainField2(), "Wednesday");
-
-		// test tags
-		tags = assembledShow.getMetadataTags();
-		tagsList = new ArrayList<>();
-		List<MetadataType> tagsList2 = new ArrayList<>();
-		tagsList.add(MetadataType.HUMIDITY);
-		tagsList2.add(MetadataType.CURRENT_TEMPERATURE);
-		assertEquals(tags.getMainField1(), tagsList);
-		assertEquals(tags.getMainField2(), tagsList2);
-
-		textAndGraphicManager.setTextField3("My");
-		textAndGraphicManager.setTextField3Type(MetadataType.MEDIA_ALBUM);
-
-		assembledShow = textAndGraphicManager.assembleShowText(inputShow);
-		assertEquals(assembledShow.getMainField1(), "It is - Wednesday");
-		assertEquals(assembledShow.getMainField2(), "My");
-
-		// test tags
-		tags = assembledShow.getMetadataTags();
-		tagsList = new ArrayList<>();
-		tagsList2 = new ArrayList<>();
-		tagsList.add(MetadataType.CURRENT_TEMPERATURE);
-		tagsList.add(MetadataType.HUMIDITY);
-		tagsList2.add(MetadataType.MEDIA_ALBUM);
-		assertEquals(tags.getMainField1(), tagsList);
-		assertEquals(tags.getMainField2(), tagsList2);
-
-		textAndGraphicManager.setTextField4("Dudes");
-		textAndGraphicManager.setTextField4Type(MetadataType.MEDIA_STATION);
-
-		assembledShow = textAndGraphicManager.assembleShowText(inputShow);
-		assertEquals(assembledShow.getMainField1(), "It is - Wednesday");
-		assertEquals(assembledShow.getMainField2(), "My - Dudes");
-
-		// test tags
-		tags = assembledShow.getMetadataTags();
-		tagsList = new ArrayList<>();
-		tagsList2 = new ArrayList<>();
-		tagsList.add(MetadataType.CURRENT_TEMPERATURE);
-		tagsList.add(MetadataType.HUMIDITY);
-		tagsList2.add(MetadataType.MEDIA_STATION);
-		tagsList2.add(MetadataType.MEDIA_ALBUM);
-		assertEquals(tags.getMainField1(), tagsList);
-		assertEquals(tags.getMainField2(), tagsList2);
-
-		// For some obscurity, lets try setting just fields 2 and 4 for a 2 line display
-		textAndGraphicManager.setTextField1(null);
-		textAndGraphicManager.setTextField3(null);
-		textAndGraphicManager.setTextField1Type(null);
-		textAndGraphicManager.setTextField3Type(null);
-
-		assembledShow = textAndGraphicManager.assembleShowText(inputShow);
-		assertEquals(assembledShow.getMainField1(), "Wednesday");
-		assertEquals(assembledShow.getMainField2(), "Dudes");
-
-		// And 3 fields without setting 1
-		textAndGraphicManager.setTextField3("My");
-
-		assembledShow = textAndGraphicManager.assembleShowText(inputShow);
-		assertEquals(assembledShow.getMainField1(), "Wednesday");
-		assertEquals(assembledShow.getMainField2(), "My - Dudes");
-
-		// test tags
-		tags = assembledShow.getMetadataTags();
-		tagsList = new ArrayList<>();
-		tagsList2 = new ArrayList<>();
-		tagsList.add(MetadataType.CURRENT_TEMPERATURE);
-		tagsList2.add(MetadataType.MEDIA_STATION);
-		assertEquals(tags.getMainField1(), tagsList);
-		assertEquals(tags.getMainField2(), tagsList2);
-	}
-
-	@Test
-	public void testAssemble3Lines() {
-
-		Show inputShow = new Show();
-
-		// Force it to return display with support for only 3 lines of text
-		textAndGraphicManager.defaultMainWindowCapability = getWindowCapability(3);
-
-		textAndGraphicManager.setTextField1("It is");
-		textAndGraphicManager.setTextField1Type(MetadataType.HUMIDITY);
-
-		Show assembledShow = textAndGraphicManager.assembleShowText(inputShow);
-		assertEquals(assembledShow.getMainField1(), "It is");
-		assertEquals(assembledShow.getMainField2(), "");
-		assertEquals(assembledShow.getMainField3(), "");
-
-		// test tags
-		MetadataTags tags = assembledShow.getMetadataTags();
-		List<MetadataType> tagsList = new ArrayList<>();
-		tagsList.add(MetadataType.HUMIDITY);
-		assertEquals(tags.getMainField1(), tagsList);
-
-		textAndGraphicManager.setTextField2("Wednesday");
-		textAndGraphicManager.setTextField2Type(MetadataType.CURRENT_TEMPERATURE);
-
-		assembledShow = textAndGraphicManager.assembleShowText(inputShow);
-		assertEquals(assembledShow.getMainField1(), "It is");
-		assertEquals(assembledShow.getMainField2(), "Wednesday");
-		assertEquals(assembledShow.getMainField3(), "");
-
-		// test tags
-		tags = assembledShow.getMetadataTags();
-		tagsList = new ArrayList<>();
-		List<MetadataType> tagsList2 = new ArrayList<>();
-		tagsList.add(MetadataType.HUMIDITY);
-		tagsList2.add(MetadataType.CURRENT_TEMPERATURE);
-		assertEquals(tags.getMainField1(), tagsList);
-		assertEquals(tags.getMainField2(), tagsList2);
-
-		textAndGraphicManager.setTextField3("My");
-		textAndGraphicManager.setTextField3Type(MetadataType.MEDIA_ALBUM);
-
-		assembledShow = textAndGraphicManager.assembleShowText(inputShow);
-		assertEquals(assembledShow.getMainField1(), "It is");
-		assertEquals(assembledShow.getMainField2(), "Wednesday");
-		assertEquals(assembledShow.getMainField3(), "My");
-
-		// test tags
-		tags = assembledShow.getMetadataTags();
-		tagsList = new ArrayList<>();
-		tagsList2 = new ArrayList<>();
-		List<MetadataType> tagsList3 = new ArrayList<>();
-		tagsList.add(MetadataType.HUMIDITY);
-		tagsList2.add(MetadataType.CURRENT_TEMPERATURE);
-		tagsList3.add(MetadataType.MEDIA_ALBUM);
-		assertEquals(tags.getMainField1(), tagsList);
-		assertEquals(tags.getMainField2(), tagsList2);
-		assertEquals(tags.getMainField3(), tagsList3);
-
-		textAndGraphicManager.setTextField4("Dudes");
-		textAndGraphicManager.setTextField4Type(MetadataType.MEDIA_STATION);
-
-		assembledShow = textAndGraphicManager.assembleShowText(inputShow);
-		assertEquals(assembledShow.getMainField1(), "It is");
-		assertEquals(assembledShow.getMainField2(), "Wednesday");
-		assertEquals(assembledShow.getMainField3(), "My - Dudes");
-
-		// test tags
-		tags = assembledShow.getMetadataTags();
-		tagsList = new ArrayList<>();
-		tagsList2 = new ArrayList<>();
-		tagsList3 = new ArrayList<>();
-		tagsList.add(MetadataType.HUMIDITY);
-		tagsList2.add(MetadataType.CURRENT_TEMPERATURE);
-		tagsList3.add(MetadataType.MEDIA_ALBUM);
-		tagsList3.add(MetadataType.MEDIA_STATION);
-		assertEquals(tags.getMainField1(), tagsList);
-		assertEquals(tags.getMainField2(), tagsList2);
-		assertEquals(tags.getMainField3(), tagsList3);
-
-		// Someone might not want to set the fields in order? We should handle that
-		textAndGraphicManager.setTextField1(null);
-
-		assembledShow = textAndGraphicManager.assembleShowText(inputShow);
-		try {
-			System.out.println(assembledShow.serializeJSON().toString());
-		} catch (JSONException e) {
-			e.printStackTrace();
-		}
-
-		assertEquals(assembledShow.getMainField2(), "Wednesday");
-		assertEquals(assembledShow.getMainField3(), "My - Dudes");
-	}
-
-	@Test
-	public void testAssemble4Lines() {
-
-		Show inputShow = new Show();
-
-		textAndGraphicManager.defaultMainWindowCapability = getWindowCapability(4);
-		TextField tx1 = new TextField();
-		TextField tx2 = new TextField();
-		TextField tx3 = new TextField();
-		TextField tx4 = new TextField();
-		TextField tx5 = new TextField();
-		TextField tx6 = new TextField();
-
-		tx1.setName(TextFieldName.mainField1);
-		tx2.setName(TextFieldName.mainField2);
-		tx3.setName(TextFieldName.mainField3);
-		tx4.setName(TextFieldName.mainField4);
-		tx5.setName(TextFieldName.mediaTrack);
-		tx6.setName(TextFieldName.templateTitle);
-
-		List<TextField> textFieldNames = Arrays.asList(tx1,tx2,tx3,tx4,tx5,tx6);
-		textAndGraphicManager.defaultMainWindowCapability.setTextFields(textFieldNames);
-
-		textAndGraphicManager.setMediaTrackTextField("HI");
-		textAndGraphicManager.setTitle("bye");
-
-		// Force it to return display with support for only 4 lines of text
-
-		textAndGraphicManager.setTextField1("It is");
-		textAndGraphicManager.setTextField1Type(MetadataType.HUMIDITY);
-
-		Show assembledShow = textAndGraphicManager.assembleShowText(inputShow);
-
-		assertEquals(assembledShow.getMainField1(), "It is");
-		assertEquals(assembledShow.getMainField2(), "");
-		assertEquals(assembledShow.getMainField3(), "");
-		assertEquals(assembledShow.getMainField4(), "");
-		assertEquals(assembledShow.getMediaTrack(), "HI");
-		assertEquals(assembledShow.getTemplateTitle(), "bye");
-
-		// test tags
-		MetadataTags tags = assembledShow.getMetadataTags();
-		List<MetadataType> tagsList = new ArrayList<>();
-		tagsList.add(MetadataType.HUMIDITY);
-		assertEquals(tags.getMainField1(), tagsList);
-
-		textAndGraphicManager.setTextField2("Wednesday");
-		textAndGraphicManager.setTextField2Type(MetadataType.CURRENT_TEMPERATURE);
-
-		assembledShow = textAndGraphicManager.assembleShowText(inputShow);
-		assertEquals(assembledShow.getMainField1(), "It is");
-		assertEquals(assembledShow.getMainField2(), "Wednesday");
-		assertEquals(assembledShow.getMainField3(), "");
-		assertEquals(assembledShow.getMainField4(), "");
-
-		// test tags
-		tags = assembledShow.getMetadataTags();
-		tagsList = new ArrayList<>();
-		List<MetadataType> tagsList2 = new ArrayList<>();
-		tagsList.add(MetadataType.HUMIDITY);
-		tagsList2.add(MetadataType.CURRENT_TEMPERATURE);
-		assertEquals(tags.getMainField1(), tagsList);
-		assertEquals(tags.getMainField2(), tagsList2);
-
-		textAndGraphicManager.setTextField3("My");
-		textAndGraphicManager.setTextField3Type(MetadataType.MEDIA_ALBUM);
-
-		assembledShow = textAndGraphicManager.assembleShowText(inputShow);
-		assertEquals(assembledShow.getMainField1(), "It is");
-		assertEquals(assembledShow.getMainField2(), "Wednesday");
-		assertEquals(assembledShow.getMainField3(), "My");
-		assertEquals(assembledShow.getMainField4(), "");
-
-		// test tags
-		tags = assembledShow.getMetadataTags();
-		tagsList = new ArrayList<>();
-		tagsList2 = new ArrayList<>();
-		List<MetadataType> tagsList3 = new ArrayList<>();
-		tagsList.add(MetadataType.HUMIDITY);
-		tagsList2.add(MetadataType.CURRENT_TEMPERATURE);
-		tagsList3.add(MetadataType.MEDIA_ALBUM);
-		assertEquals(tags.getMainField1(), tagsList);
-		assertEquals(tags.getMainField2(), tagsList2);
-		assertEquals(tags.getMainField3(), tagsList3);
-
-		textAndGraphicManager.setTextField4("Dudes");
-		textAndGraphicManager.setTextField4Type(MetadataType.MEDIA_STATION);
-
-		assembledShow = textAndGraphicManager.assembleShowText(inputShow);
-		assertEquals(assembledShow.getMainField1(), "It is");
-		assertEquals(assembledShow.getMainField2(), "Wednesday");
-		assertEquals(assembledShow.getMainField3(), "My");
-		assertEquals(assembledShow.getMainField4(), "Dudes");
-
-		// test tags
-		tags = assembledShow.getMetadataTags();
-		tagsList = new ArrayList<>();
-		tagsList2 = new ArrayList<>();
-		tagsList3 = new ArrayList<>();
-		List<MetadataType> tagsList4 = new ArrayList<>();
-		tagsList.add(MetadataType.HUMIDITY);
-		tagsList2.add(MetadataType.CURRENT_TEMPERATURE);
-		tagsList3.add(MetadataType.MEDIA_ALBUM);
-		tagsList4.add(MetadataType.MEDIA_STATION);
-		assertEquals(tags.getMainField1(), tagsList);
-		assertEquals(tags.getMainField2(), tagsList2);
-		assertEquals(tags.getMainField3(), tagsList3);
-		assertEquals(tags.getMainField4(), tagsList4);
-
-		// try just setting line 1 and 4
-		textAndGraphicManager.setTextField2(null);
-		textAndGraphicManager.setTextField3(null);
-		textAndGraphicManager.setTextField2Type(null);
-		textAndGraphicManager.setTextField3Type(null);
-
-		assembledShow = textAndGraphicManager.assembleShowText(inputShow);
-		assertEquals(assembledShow.getMainField1(), "It is");
-		assertEquals(assembledShow.getMainField2(), "");
-		assertEquals(assembledShow.getMainField3(), "");
-		assertEquals(assembledShow.getMainField4(), "Dudes");
-
-		// test tags
-		tags = assembledShow.getMetadataTags();
-		tagsList = new ArrayList<>();
-		tagsList4 = new ArrayList<>();
-		tagsList.add(MetadataType.HUMIDITY);
-		tagsList4.add(MetadataType.MEDIA_STATION);
-		assertEquals(tags.getMainField1(), tagsList);
-		assertEquals(tags.getMainField4(), tagsList4);
-	}
-
-	*//**
-	 * Testing if WindowCapability is null, TextFields should still update.
-	 *//*
-	@Test
-	public void testAssemble4LinesNullWindowCapability() {
-
-		Show inputShow = new Show();
-		
-		textAndGraphicManager.setMediaTrackTextField("HI");
-		textAndGraphicManager.setTitle("bye");
-
-		textAndGraphicManager.setTextField1("It is");
-		textAndGraphicManager.setTextField1Type(MetadataType.HUMIDITY);
-
-		Show assembledShow = textAndGraphicManager.assembleShowText(inputShow);
-
-		assertEquals(assembledShow.getMainField1(), "It is");
-		assertEquals(assembledShow.getMainField2(), "");
-		assertEquals(assembledShow.getMainField3(), "");
-		assertEquals(assembledShow.getMainField4(), "");
-		assertEquals(assembledShow.getMediaTrack(), "HI");
-		assertEquals(assembledShow.getTemplateTitle(), "bye");
-
-		// test tags
-		MetadataTags tags = assembledShow.getMetadataTags();
-		List<MetadataType> tagsList = new ArrayList<>();
-		tagsList.add(MetadataType.HUMIDITY);
-		assertEquals(tags.getMainField1(), tagsList);
-
-		textAndGraphicManager.setTextField2("Wednesday");
-		textAndGraphicManager.setTextField2Type(MetadataType.CURRENT_TEMPERATURE);
-
-		assembledShow = textAndGraphicManager.assembleShowText(inputShow);
-		assertEquals(assembledShow.getMainField1(), "It is");
-		assertEquals(assembledShow.getMainField2(), "Wednesday");
-		assertEquals(assembledShow.getMainField3(), "");
-		assertEquals(assembledShow.getMainField4(), "");
-
-		// test tags
-		tags = assembledShow.getMetadataTags();
-		tagsList = new ArrayList<>();
-		List<MetadataType> tagsList2 = new ArrayList<>();
-		tagsList.add(MetadataType.HUMIDITY);
-		tagsList2.add(MetadataType.CURRENT_TEMPERATURE);
-		assertEquals(tags.getMainField1(), tagsList);
-		assertEquals(tags.getMainField2(), tagsList2);
-
-		textAndGraphicManager.setTextField3("My");
-		textAndGraphicManager.setTextField3Type(MetadataType.MEDIA_ALBUM);
-
-		assembledShow = textAndGraphicManager.assembleShowText(inputShow);
-		assertEquals(assembledShow.getMainField1(), "It is");
-		assertEquals(assembledShow.getMainField2(), "Wednesday");
-		assertEquals(assembledShow.getMainField3(), "My");
-		assertEquals(assembledShow.getMainField4(), "");
-
-		// test tags
-		tags = assembledShow.getMetadataTags();
-		tagsList = new ArrayList<>();
-		tagsList2 = new ArrayList<>();
-		List<MetadataType> tagsList3 = new ArrayList<>();
-		tagsList.add(MetadataType.HUMIDITY);
-		tagsList2.add(MetadataType.CURRENT_TEMPERATURE);
-		tagsList3.add(MetadataType.MEDIA_ALBUM);
-		assertEquals(tags.getMainField1(), tagsList);
-		assertEquals(tags.getMainField2(), tagsList2);
-		assertEquals(tags.getMainField3(), tagsList3);
-
-		textAndGraphicManager.setTextField4("Dudes");
-		textAndGraphicManager.setTextField4Type(MetadataType.MEDIA_STATION);
-
-		assembledShow = textAndGraphicManager.assembleShowText(inputShow);
-		assertEquals(assembledShow.getMainField1(), "It is");
-		assertEquals(assembledShow.getMainField2(), "Wednesday");
-		assertEquals(assembledShow.getMainField3(), "My");
-		assertEquals(assembledShow.getMainField4(), "Dudes");
-
-		// test tags
-		tags = assembledShow.getMetadataTags();
-		tagsList = new ArrayList<>();
-		tagsList2 = new ArrayList<>();
-		tagsList3 = new ArrayList<>();
-		List<MetadataType> tagsList4 = new ArrayList<>();
-		tagsList.add(MetadataType.HUMIDITY);
-		tagsList2.add(MetadataType.CURRENT_TEMPERATURE);
-		tagsList3.add(MetadataType.MEDIA_ALBUM);
-		tagsList4.add(MetadataType.MEDIA_STATION);
-		assertEquals(tags.getMainField1(), tagsList);
-		assertEquals(tags.getMainField2(), tagsList2);
-		assertEquals(tags.getMainField3(), tagsList3);
-		assertEquals(tags.getMainField4(), tagsList4);
-
-		// try just setting line 1 and 4
-		textAndGraphicManager.setTextField2(null);
-		textAndGraphicManager.setTextField3(null);
-		textAndGraphicManager.setTextField2Type(null);
-		textAndGraphicManager.setTextField3Type(null);
-
-		assembledShow = textAndGraphicManager.assembleShowText(inputShow);
-		assertEquals(assembledShow.getMainField1(), "It is");
-		assertEquals(assembledShow.getMainField2(), "");
-		assertEquals(assembledShow.getMainField3(), "");
-		assertEquals(assembledShow.getMainField4(), "Dudes");
-
-		// test tags
-		tags = assembledShow.getMetadataTags();
-		tagsList = new ArrayList<>();
-		tagsList4 = new ArrayList<>();
-		tagsList.add(MetadataType.HUMIDITY);
-		tagsList4.add(MetadataType.MEDIA_STATION);
-		assertEquals(tags.getMainField1(), tagsList);
-		assertEquals(tags.getMainField4(), tagsList4);
-	}*/
-
 	@Test
 	public void testMediaTrackTextField() {
 
@@ -664,22 +176,6 @@ public class TextAndGraphicManagerTests {
 		textAndGraphicManager.setTextAlignment(TextAlignment.LEFT_ALIGNED);
 		assertEquals(textAndGraphicManager.getTextAlignment(), TextAlignment.LEFT_ALIGNED);
 	}
-
-/*	@Test
-	public void testExtractTextFromShow(){
-
-		Show mainShow = new Show();
-		mainShow.setMainField1("test");
-		mainShow.setMainField3("Sauce");
-		mainShow.setMainField4("");
-
-		Show newShow = textAndGraphicManager.extractTextFromShow(mainShow);
-
-		assertEquals(newShow.getMainField1(), "test");
-		assertEquals(newShow.getMainField3(), "Sauce");
-		assertEquals(newShow.getMainField4(), "");
-		assertNull(newShow.getMainField2());
-	}*/
 
 	// TEST IMAGES
